Controller Veiculo com os recursos listar, cadastrar e editar veiculos;
Mudança atributo Id de Integer para Long na classe DAO<T>;
Métodos alterar, busca e busca por placa na classe VeiculoDAO;
Exclusão de anotações não usadas;
Menu principal (provisório) no layout;
Página lista para o recurso listar veiculos;
Página formulário para os recursos cadastras e editar veiculos;
Validar se já existe um veiculo cadastrado com determinada placa;

+++ Gerenciamento-de-Frota/src/main/java/gerenciamentodefrota/dao/DAO.java
@@ -15,7 +15,7 @@ public class DAO<T> {
 		this.classe = classe;
 	}
 
-	public T busca(Integer id) {
+	public T busca(Long id) {
 		return em.getReference(classe, id);
 	}
